@@ -28,52 +28,51 @@
     "serve": "cd tools && npm run cli:serve"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.22",
-    "@angular-devkit/build-ng-packagr": "^0.803.22",
-    "@angular/animations": "^8.2.14",
-    "@angular/cdk": "^8.2.3",
-    "@angular/cli": "^8.3.22",
-    "@angular/common": "^8.2.14",
-    "@angular/compiler": "^8.2.14",
-    "@angular/compiler-cli": "^8.2.14",
-    "@angular/core": "^8.2.14",
-    "@angular/flex-layout": "^8.0.0-beta.27",
-    "@angular/forms": "^8.2.14",
+    "@angular-devkit/build-angular": "~12.2.16",
+    "@angular-devkit/core": "^12.2.16",
+    "@angular/animations": "^12.2.16",
+    "@angular/cdk": "^12.2.13",
+    "@angular/cli": "^12.2.16",
+    "@angular/common": "^12.2.16",
+    "@angular/compiler": "^12.2.16",
+    "@angular/compiler-cli": "^12.2.16",
+    "@angular/core": "^12.2.16",
+    "@angular/flex-layout": "^12.0.0-beta.35",
+    "@angular/forms": "^12.2.16",
     "@angular/http": "^7.2.16",
-    "@angular/material": "^8.2.3",
-    "@angular/platform-browser": "^8.2.14",
-    "@angular/platform-browser-dynamic": "^8.2.14",
-    "@angular/router": "^8.2.14",
+    "@angular/material": "^12.2.13",
+    "@angular/platform-browser": "^12.2.16",
+    "@angular/platform-browser-dynamic": "^12.2.16",
+    "@angular/router": "^12.2.16",
     "@compodoc/compodoc": "^1.1.11",
-    "@firestitch/api": "^8.1.4",
-    "@firestitch/common": "^7.0.5",
-    "@firestitch/drawer": "^8.0.2",
-    "@firestitch/example": "^8.0.0",
-    "@firestitch/iframe": "^8.0.0",
-    "@firestitch/menu": "^7.0.11",
-    "@firestitch/message": "^8.1.0",
+    "@firestitch/api": "^12.0.0",
+    "@firestitch/common": "^12.0.0",
+    "@firestitch/drawer": "^12.0.0",
+    "@firestitch/example": "^12.0.0",
+    "@firestitch/iframe": "^12.0.0",
+    "@firestitch/menu": "^12.0.1",
+    "@firestitch/message": "^12.0.0",
     "@ngx-loading-bar/core": "^4.2.0",
     "@ngx-loading-bar/http-client": "^4.2.0",
-    "codelyzer": "^5.2.1",
+    "codelyzer": "^6.0.0",
     "cpx": "^1.5.0",
     "date-fns": "^2.9.0",
     "hammerjs": "^2.0.8",
     "highlight.js": "^9.17.1",
     "lodash-es": "^4.17.15",
-    "ng-packagr": "^5.7.1",
-    "ngx-toastr": "^11.3.0",
+    "ng-packagr": "^12.2.7",
+    "ngx-highlightjs": "^6.1.2",
+    "ngx-toastr": "^14.2.2",
     "npm-run-all": "^4.1.5",
     "rxjs": "^6.5.4",
-    "tsickle": "^0.38.0",
-    "tslint": "^5.20.1",
-    "typescript": "^3.5.3",
-    "webpack": "^4.41.5",
-    "webpack-dev-server": "^3.10.1",
-    "zone.js": "^0.10.2"
+    "tslint": "~6.1.0",
+    "typescript": "4.3.5",
+    "zone.js": "~0.11.4"
   },
   "engines": {
     "node": ">=6.0.0"
   },
-  "peerDependencies": {},
-  "dependencies": {}
+  "dependencies": {
+    "tslib": "^2.0.0"
+  }
 }
